+++ spec/support/named_internal_node.rb
@@ -1,0 +1,11 @@
+class NamedInternalNode
+  include CallableTree::Node::Internal
+
+  def initialize(name)
+    @name = name
+  end
+
+  def identity
+    @name
+  end
+end